python-sdk-resource-creation-samples/create-resource-group/hosted/requirements.txt
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
flask